reduced spin_rate and now the reachability map is only read once.
Drawback of reading reachability map once:

You cannot switch to another reach map in the middle of the program.
Hence the benginning and ending of remove_obstacles reachability is not
independent of the reach map loading.

Supposed benefit:

System should no longer crash due to heavy load. Though the load may
also be due to the simulation i.e.V-rep or gazebo